src/requirements.txt
Flask==3.0.3
kafka_python==2.0.2
Requests==2.31.0
